+++ PipingServer.NET/Streams/CloseRegisterStream.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Piping
+namespace Piping.Streams
 {
     public class CloseRegisterStream : Stream
     {
