+++ PipingServer.NETTests/Streams/PipingStreamTests.cs
@@ -4,13 +4,11 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Piping;
 using static DebugUtils;
 
-namespace Piping.Tests
+namespace Piping.Streams.Tests
 {
     [TestClass]
     public class PipingStreamTests
